requirement.txt
Django==3.2.16
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.1.0
Pillow==4.3.0
psycopg2==2.9.5
python-decouple==3.3
urllib3==1.24.1
whitenoise==5.3.0

